Complete build functionality. Used mwarning's server for data.
Now users can build their images using the Build button and it works
fine for the most part.
There is still some issue with the file download as there are a lot
of them and will be imporved in the future.
Using axios now for cross browser support issue of `fetch`

Signed-off-by: Sudhanshu Gautam <me@sudhanshug.com>